Squashed commit of the following:
commit 689a8e701525863d6b78172145a701b2831afae9
Author: LiDong <cnlidong@live.cn>
Date:   Sat Sep 10 22:09:18 2022 +0800

    增加模块默认值

commit be65bdabaf3862ef3e5cf3fe0ef568e0802d6d61
Author: LiDong <cnlidong@live.cn>
Date:   Mon Sep 5 22:04:16 2022 +0800

    完善角色相关

commit aabfe25515dc79664afa9e8d245adc5852818630
Author: LiDong <cnlidong@live.cn>
Date:   Mon Sep 5 16:49:17 2022 +0800

    id change to uint

commit 3e2a8347cf33cd79e07367ab4c31817525b2c849
Author: LiDong <cnlidong@live.cn>
Date:   Sun Sep 4 18:11:54 2022 +0800

    暂存

commit 9fff79de4ca012e77d9cf6c8cda537c173313fc7
Author: LiDong <cnlidong@live.cn>
Date:   Sun Sep 4 17:54:02 2022 +0800

    暂存

commit 86fcd234cb920e38db24f14fa1d0e630c27f16cb
Author: LiDong <cnlidong@live.cn>
Date:   Sun Sep 4 12:07:01 2022 +0800

    完成space模块

commit 200cbff4ded688a00334303ead0e86dc3d3c3786
Author: LiDong <cnlidong@live.cn>
Date:   Sat Sep 3 23:48:45 2022 +0800

    暂存

+++ modules/base/model/base_sys_role_department.go
@@ -7,8 +7,8 @@ const TableNameBaseSysRoleDepartment = "base_sys_role_department"
 // BaseSysRoleDepartment mapped from table <base_sys_role_department>
 type BaseSysRoleDepartment struct {
 	*cool.Model
-	RoleID       int64 `gorm:"column:roleId;type:bigint;not null" json:"roleId"`             // 角色ID
-	DepartmentID int64 `gorm:"column:departmentId;type:bigint;not null" json:"departmentId"` // 部门ID
+	RoleID       uint `gorm:"column:roleId;type:bigint;not null" json:"roleId"`             // 角色ID
+	DepartmentID uint `gorm:"column:departmentId;type:bigint;not null" json:"departmentId"` // 部门ID
 }
 
 // TableName BaseSysRoleDepartment's table name
